requirements.txt
torch
transformers[torch]
pillow
fastapi
"uvicorn[standard]"
io